pyproject.toml
[project]
name = "tts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiomysql>=0.3.2",
    "argon2-cffi>=25.1.0",
    "boto3>=1.42.74",
    "cryptography>=46.0.5",
    "detoxify>=0.5.2",
    "dotenv>=0.9.9",
    "fastapi>=0.135.2",
    "faster-whisper>=1.2.1",
    "numpy>=2.2.6",
    "passlib>=1.7.4",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "pymysql>=1.1.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.22",
    "scipy>=1.15.3",
    "soundfile>=0.13.1",
    "sqlmodel>=0.0.37",
    "torch>=2.11.0",
    "transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
